scripts/requirements.txt
requests
google-cloud-storage
python-arango==5.4.0
